+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Adminer dev
+- Do not edit NULL values by Modify (bug #967)
 - MySQL: Display number of found rows in group queries (regression from 5.1.1)
 
 ## Adminer 5.1.1 (released 2025-04-02)